Add files via upload
I needed to create a tool (an analysis notebook) that analyzes and visualizes the major metrics of the portfolios across all of these areas, and determine which portfolio outperformed the others. I was given the historical daily returns of several portfolios: some from the firm's algorithmic portfolios, some that represent the portfolios of famous "whale" investors like Warren Buffett, and some from the big hedge and mutual funds. Then I used this analysis to create a custom portfolio of stocks and compared its performance to that of the other portfolios, as well as the larger market (S&P 500 Index).